加入互信息过滤,尝试更换归一化方法,并且在打分计算中加入左右信息熵紧密度 这一统计信息

+++ src/main/java/config/Constants.java
@@ -25,9 +25,9 @@ public class Constants {
 
     public static Map<String, Term> segTermMap = new HashMap<>();   //用于统计量归一化计算
 
-    public static final List<Double> LElist = new ArrayList<>();   //左熵
-    public static final List<Double> RElist = new ArrayList<>();   //右熵
-    public static final List<Double> MI_list = new ArrayList<>();   //互信息
+    public static final List<Float> LElist = new ArrayList<>();   //左熵
+    public static final List<Float> RElist = new ArrayList<>();   //右熵
+    public static final List<Float> MI_list = new ArrayList<>();   //互信息
 
     /**
      * 字典树,用于左前缀查询
